Load 3.3.37 into branch/prcs-import-33.
git-svn-id: svn://svn.bearnip.com/ldmud/branch/prcs-import-33@127 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/parse_old.c
@@ -262,7 +262,7 @@ find_living_object (const char *name, Bool player)
     put_c_string(sp, name);
     if (!sp->u.str)
         error("(parse_command) Out of memory (%lu bytes) for result\n"
-             , strlen(name));
+             , (unsigned long) strlen(name));
     inter_sp = sp;
     call_lambda(svp, 1);
     pop_stack();
